feat: change cookieAdapter const to createCookieAdapter function to be able to pass custom sessionKey
- minor function extraction for better readability
