+++ src/session-adapters/index.ts
@@ -1,3 +1,3 @@
-export * from './cookieAdapter'
+export * from './createCookieAdapter'
 export * from './publicAdapter'
 export * from './internalAdapter'
